reorganzie file structure
