+++ identity-service/src/typed-routes/solana/programs/claimable-tokens/constants.ts
@@ -1,0 +1,10 @@
+import { PublicKey } from '@solana/web3.js'
+
+export enum ClaimableTokenInstruction {
+  Create = 0,
+  Transfer = 1
+}
+
+export const CLAIMABLE_TOKEN_PROGRAM_ID = new PublicKey(
+  'Ewkv3JahEFRKkcJmpoKB7pXbnUHwjAyXiwEo4ZY2rezQ'
+)
