+++ identity-service/src/typed-routes/solana/programs/reward-manager/index.ts
@@ -1,0 +1,5 @@
+export * from './decode'
+export * from './createSenderPublic'
+export * from './deleteSenderPublic'
+export * from './evaluateAttestations'
+export * from './submitAttestation'
